internal/update: removed obsolete code & refactored
