+++ internal/update/update.go
@@ -2,19 +2,74 @@ package update
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strings"
+	"time"
 
-	"github.com/blang/semver"
 	"github.com/cli/safeexec"
-	"github.com/superfly/flyctl/terminal"
+
+	"github.com/superfly/flyctl/internal/buildinfo"
+	"github.com/superfly/flyctl/internal/cmdutil"
+	"github.com/superfly/flyctl/internal/env"
+	"github.com/superfly/flyctl/pkg/iostreams"
 )
 
-// Check whether the fly binary was found under the Homebrew prefix
+type Release struct {
+	Version     string    `yaml:"version"`
+	Prerelease  bool      `yaml:"prerelease"`
+	DownloadURL string    `yaml:"download_url" json:"download_url"`
+	Timestamp   time.Time `yaml:"timestamp"`
+}
+
+// Check reports whether update checks should take place.
+func Check() bool {
+	switch {
+	case env.IsTruthy("FLY_UPDATE_CHECK"):
+		return true
+	case env.IsTruthy("FLY_NO_UPDATE_CHECK"):
+		return false
+	case env.IsSet("CODESPACES"):
+		return false
+	case !buildinfo.IsRelease(), env.IsCI():
+		return false
+	case !cmdutil.IsTerminal(os.Stdout), !cmdutil.IsTerminal(os.Stderr):
+		return false
+	default:
+		return true
+	}
+}
+
+// LatestRelease reports the latest release for the given channel.
+func LatestRelease(ctx context.Context, channel string) (*Release, error) {
+	updateUrl := fmt.Sprintf("https://api.fly.io/app/flyctl_releases/%s/%s/%s", runtime.GOOS, runtime.GOARCH, channel)
+	req, err := http.NewRequestWithContext(ctx, "GET", updateUrl, nil)
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Add("Accept", "application/json")
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	var release Release
+	if err := json.NewDecoder(resp.Body).Decode(&release); err != nil {
+		return &release, err
+	}
+
+	return &release, nil
+}
+
+// isUnderHomebrew reports whether the fly binary was found under the Homebrew
+// prefix.
 func isUnderHomebrew() bool {
 	flyBinary, err := os.Executable()
 	if err != nil {
@@ -55,33 +110,7 @@ func updateCommand(prerelease bool) string {
 	}
 }
 
-func PerformInPlaceUpgrade(ctx context.Context, configPath string, currentVersion semver.Version) error {
-	state, _ := loadState(configPath)
-	if state.Channel == "" {
-		state.Channel = "latest"
-	}
-
-	release, err := fetchLatestVersion(ctx, state.Channel)
-	if err != nil {
-		return err
-	}
-
-	var latestVersion semver.Version
-	if release != nil {
-		latestVersion, err = semver.ParseTolerant(release.Version)
-		if err != nil {
-			terminal.Warnf("error parsing version number '%s': %s\n", state.LatestRelease.Version, err)
-		}
-		// if latestVersion.GT(currentVersion) {
-		// 	return state.LatestRelease, nil
-		// }
-	}
-
-	if release == nil || !latestVersion.GT(currentVersion) {
-		fmt.Println("No update available")
-		return nil
-	}
-
+func UpgradeInPlace(ctx context.Context, io *iostreams.IOStreams, prelease bool) error {
 	if runtime.GOOS == "windows" {
 		if err := renameCurrentBinaries(); err != nil {
 			return err
@@ -101,14 +130,14 @@ func PerformInPlaceUpgrade(ctx context.Context, configPath string, currentVersio
 	}
 	fmt.Println(shellToUse, switchToUse)
 
-	command := updateCommand(release.Prerelease)
+	command := updateCommand(prelease)
 
-	fmt.Println("Running automatic update [" + command + "]")
+	fmt.Fprintf(io.ErrOut, "Running automatic update [%s]\n", command)
+
 	cmd := exec.Command(shellToUse, switchToUse, command)
-
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdin
+	cmd.Stdout = io.Out
+	cmd.Stderr = io.ErrOut
+	cmd.Stdin = io.In
 
 	return cmd.Run()
 }
@@ -124,23 +153,6 @@ func renameCurrentBinaries() error {
 		if err := os.Rename(p, p+".old"); err != nil {
 			return err
 		}
-	}
-
-	return nil
-}
-
-func PostUpgradeCleanup() error {
-	if runtime.GOOS != "windows" {
-		return nil
-	}
-
-	binaries, err := currentWindowsBinaries()
-	if err != nil {
-		return err
-	}
-
-	for _, p := range binaries {
-		os.Remove(p + ".old")
 	}
 
 	return nil
